modified:   _init_.py 	modified:   azi_play.py 	modified:   plugins/zplay/deal_event.py 	renamed:    video_update.py -> plugins/zplay/video_update.py
