+++ _init_.py
@@ -3,4 +3,4 @@ from .plugins.zplay.obs_websocket import OBS_PLAY
 from .plugins.zplay.deal_event import zplay
 from .plugins.zplay.search import Search
 from .plugins.zplay.sql_redis import insert_redis
-from .video_update import download_bili
+from .plugins.zplay.video_update import download_bili
